# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task company -echo- API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "CategoryImage"
                ],
                "summary": "Get category information",
                "description": "Get category information by id",
                "operationId": "ef336795b6aaaf8cd8c9af1a4ac4649c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CategoryImage id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/copy-images/{id}": {
            "get": {
                "tags": [
                    "CopyMiniImage"
                ],
                "summary": "Get copy-images information",
                "description": "Get copy-images information by id",
                "operationId": "d2a45dbd6c59deaaac62926ea04f7fca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CopyMiniImage id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}